+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 ### Changes
 * Enhanced attributes in tooltip, sorting them by attribute and then by operation, so it's not different on same items
+* Crit, absorption and piercing attributes are now hidden from tooltips if the relative feature is disabled
 
 ## 4.21.3
 ### Hotfix